Remove tpmconfig Sysinv APIs
With the new Certificate orchestration APIs added for Distributed Cloud,
there is no need to have separate tpmconfig-XXX APIs. In the interest of
minimizing code bloat, this commit gets rid of the SysInv APIs for
tpmconfig, the CGCTS CLI equivalents, and the REST API documentation.
Notice that as a CLI only convenience we support passing in the reserved
"tpm" keyword in place of the UUID in order to save users from first
listing to query UUID and then doing a show

Depends-On: https://review.openstack.org/578413
Change-Id: I9f0eb80b07b3b5160d7dc4eacae452f2be5a1534
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>