Enable configurable vswitch memory
Currently, a DPDK enabled vswitch makes use of a fixed 1G hugepage to
enable an optimized datapath.

In the case of OVS-DPDK, this can cause an issue when changing the
MTU of one or more interfaces, as a separate mempool is allocated
for each size.  If the minimal mempool size(s) cannot fit into the
1G page, DPDK memory initialization will fail.

This commit allows an operator to configure the amount of hugepage
memory allocated to each socket on a host, which can enable
jumboframe support for OVS-DPDK.

The system memory command has been modified to accept vswitch
hugepage configuration via the function flag. ie:

system host-memory-modify -f vswitch -1G 4 <worker_name> <node>

Story: 2004472
Task: 28162
Change-Id: Ie58d1b8317b4eb3c498524db6bd53ffff3bf1a39
Signed-off-by: Steven Webster <steven.webster@windriver.com>