Remove sdk package from cgts-client.spec
A previous update removed the "%files sdk" section from
the cgts-client.spec file, but missed removal
of the "%package sdk" section. As a result, the build
system looks for the cgts-client-sdk package
when checking if a rebuild is required. But because the
"%files" was removed, the package is not built. This leads
the build system to decide it needs to rebuild.

Change-Id: I2c293e73ad25b72c8218ba9d090ebf1c47fab44e
Closes-Bug: 1830284
Signed-off-by: Don Penney <don.penney@windriver.com>