Support adding a trusted CA certificate
We currently don't support adding a trusted CA certificate. This might
be required to connect to a docker registry using a self signed
certificate.

This change extends the certificate-install system command with the
ssl_ca mode. The certificate will be installed on all nodes via a
runtime puppet manifest.

Change-Id: I126ff67ed4682879e6ca11710503a316bd97ddf7
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Partial-Bug: 1831946